Changed directory architecture
